[TEST] added api tests (+ test data) 	Moved code out of main cpp files into stand-alone cpp files with connected hpp files (in include folder)
