+++ include/detect_breakends/aligned_segment.hpp
