+++ include/detect_breakends/bam_functions.hpp
