+++ include/junction.hpp
