+++ test/data/datasources.cmake
@@ -2,7 +2,19 @@ cmake_minimum_required (VERSION 3.8)
 
 include (cmake/app_datasources.cmake)
 
+# Get Hash-value via bash: 'shasum -a 256 path/to/file.ending
+
 # copies file to <build>/data/in.fastq
 declare_datasource (FILE in.fastq
                     URL ${CMAKE_SOURCE_DIR}/test/data/in.fastq
                     URL_HASH SHA256=6e30fc35f908a36fe0c68a7a35c47f51f9570da16622fb0c072a20e6a9ba5b3e)
+
+# copies file to <build>/data/converted_bam_shorted.sam
+declare_datasource (FILE converted_bam_shorted.sam
+                    URL ${CMAKE_SOURCE_DIR}/test/data/converted_bam_shorted.sam
+                    URL_HASH SHA256=cde4647e080dda882a0748a0ec34613b8de150a09b90cd831f85034003a9828c)
+
+# copies file to <build>/data/detect_breakends_shorted.vcf
+declare_datasource (FILE detect_breakends_shorted.vcf
+                    URL ${CMAKE_SOURCE_DIR}/test/data/detect_breakends_shorted.vcf
+                    URL_HASH SHA256=c81247543fb3603772ddd77f00e30f74d6f05e9e7f5f008ddcd516e18b040dab)
