+++ test/data/detect_breakends_shorted.vcf
@@ -1,0 +1,12 @@
+Reference	chr14	22077442	Forward	Reference	chr21	29998677	Reverse	m9232/2314/CCS
+Reference	chr21	9435236	Forward	Reference	chr21	9435286	Forward	m56874/12088/CCS
+Reference	chr21	9540623	Forward	Read 	0	7829	Forward	m8141/13515/CCS
+Reference	chr21	10177365	Forward	Reference	chr21	37992489	Forward	m15438/10156/CCS
+Reference	chr21	10177375	Forward	Reference	chr21	37992489	Forward	m50205/8540/CCS
+Reference	chr21	10398794	Forward	Read 	24	8653	Forward	m23644/13314/CCS
+Reference	chr21	10398795	Forward	Read 	25	6672	Forward	m18773/10370/CCS
+Reference	chr21	10723245	Forward	Reference	chr21	10960598	Forward	m53873/15553/CCS
+Reference	chr21	10906077	Forward	Read 	28	9408	Forward	m17204/14352/CCS
+Reference	chr21	10946747	Forward	Read 	35	2278	Forward	m13480/6038/CCS
+Reference	chr21	11069884	Forward	Reference	chr21	40076888	Forward	m14096/8850/CCS
+Reference	chr21	11104574	Forward	Reference	chr21	11104822	Forward	m43478/14786/CCS
